Feat: 헤더 분리
- 헤더에서 사용하는 함수들 Header 컴포넌트로 이동
