+++ src/components/HomePage/Home.js
@@ -1,19 +1,16 @@
-import { Button, Col, Divider, PageHeader } from 'antd';
+import { Col, Divider } from 'antd';
 import Search from 'antd/lib/input/Search';
-import { useSelector } from 'react-redux';
 import LayoutTop from './LayoutTop';
 import LayoutBottom from './LayoutBottom';
 import styles from '../../styles/Home.module.css';
-import { FaBookMedical } from 'react-icons/fa';
-import { getQuery } from '../../utils/api';
 import { useEffect, useRef, useState } from 'react';
 import TodaysWord from './TodaysWord';
 import MyTest from './MyTest';
 import Encyclopedia from './Encyclopedia';
 import NextButton from './NextButton';
+import Header from '../NavBar/Header';
 
-export default function Home({ search, login, signup, logout, mybook }) {
-  const token = useSelector((state) => state.user.auth.token);
+export default function Home({ search, login }) {
   const searchRef = useRef(null);
 
   // 슬라이드 개수(0부터 시작)
@@ -61,41 +58,7 @@ export default function Home({ search, login, signup, logout, mybook }) {
 
   return (
     <div>
-      {!token ? (
-        <PageHeader
-          title={
-            <div className={styles.header_title}>
-              <FaBookMedical className={styles.header_title_icon} />
-              <div>WORDBOOK</div>
-            </div>
-          }
-          extra={[
-            <Button key="1" onClick={login} className={styles.button}>
-              로그인
-            </Button>,
-            <Button key="2" onClick={signup} className={styles.button}>
-              회원가입
-            </Button>,
-          ]}
-        />
-      ) : (
-        <PageHeader
-          title={
-            <div className={styles.header_title}>
-              <FaBookMedical className={styles.header_title_icon} />
-              <div>WORDBOOK</div>
-            </div>
-          }
-          extra={[
-            <Button key="1" onClick={mybook} className={styles.button}>
-              마이북
-            </Button>,
-            <Button key="2" onClick={logout} className={styles.button}>
-              로그아웃
-            </Button>,
-          ]}
-        />
-      )}
+      <Header />
       <LayoutTop>
         <div className={styles.search_area}>
           <div className={styles.search_title}>[한국어 - 인도네시아어 사전]</div>
