+++ src/styles/Header.module.css
@@ -1,0 +1,23 @@
+.header_title {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.header_title_icon {
+  color: #3c3cb1;
+}
+
+.button {
+  border-color: #3c3cb1;
+  background-color: #3c3cb1;
+  text-transform: uppercase;
+  border-radius: 5px;
+  color: white;
+  font-weight: bold;
+}
+
+.button:hover {
+  background-color: #3c3cb1;
+  color: white;
+}
