+++ src/styles/Home.module.css
@@ -1,27 +1,3 @@
-.header_title {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.header_title_icon {
-  color: #3c3cb1;
-}
-
-.button {
-  border-color: #3c3cb1;
-  background-color: #3c3cb1;
-  text-transform: uppercase;
-  border-radius: 5px;
-  color: white;
-  font-weight: bold;
-}
-
-.button:hover {
-  background-color: #3c3cb1;
-  color: white;
-}
-
 .search_area {
   display: flex;
   flex-direction: column;
